apps/worker/requirements.txt
google-cloud-pubsub==2.27.1
google-cloud-storage==2.18.2
google-auth==2.36.0

psycopg2-binary==2.9.10

pillow==11.0.0




